remove 5 time duplicated function
