+++ delaunay/treeSelector.py
@@ -17,6 +17,7 @@ from PyQt4.QtGui import *
 from qgis.core import *
 from qgis.gui import *
 from processing import runalg
+from folderManager import initialize
 
 
 def main(options):
@@ -44,23 +45,6 @@ def main(options):
             filename = basename(os.path.splitext(options['filePath'])[0])
 
             # Process each file
-            processing(options, filename)
-
-
-def initialize(options):
-    '''
-    Prepare the folders for outputs:
-    '''
-    if not os.path.isdir(options['dst']):
-        os.mkdir(options['dst'])
-    if not options['dst'].endswith('/'):
-        options['dst'] = options['dst'] + '/'
-    tifdst = options['dst'] + 'tif'
-    if not os.path.exists(tifdst):
-        os.makedirs(tifdst)
-    shpdst = options['dst'] + 'shp'
-    if not os.path.exists(shpdst):
-        os.makedirs(shpdst)
 
 
 def processing(options, f):
